# Vocalis-Passport/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vocalis Passport API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /student/dashboard:
    get:
      operationId: getStudentDashboard
      tags: [student]
      summary: Get the signed-in student's dashboard
      responses:
        "200":
          description: Student dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDashboard"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /student/profile:
    get:
      operationId: getStudentProfile
      tags: [student]
      summary: Get the signed-in student's passport profile
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateStudentProfile
      tags: [student]
      summary: Update the signed-in student's passport details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /student/photo:
    post:
      operationId: uploadStudentPhoto
      tags: [student]
      summary: Upload a student portrait photo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PhotoUploadInput"
      responses:
        "200":
          description: Updated profile with photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /student/passport:
    get:
      operationId: getStudentPassport
      tags: [student]
      summary: Get the signed-in student's passport
      responses:
        "200":
          description: Passport
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passport"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: generateStudentPassport
      tags: [student]
      summary: Generate or regenerate the student's passport
      responses:
        "200":
          description: Generated passport
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /student/passport/pdf:
    get:
      operationId: downloadStudentPassportPdf
      tags: [student]
      summary: Download the signed-in student's passport as PDF
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [Storage]
      summary: Serve a private uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
components:
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Unprocessable:
      description: Passport data is incomplete
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    StudentDashboard:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/StudentProfile"
        passport:
          oneOf:
            - $ref: "#/components/schemas/Passport"
            - type: "null"
        profileCompletion:
          type: integer
          minimum: 0
          maximum: 100
      required: [student, passport, profileCompletion]
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        vocalisId:
          type: string
        profilePhotoUrl:
          type: ["string", "null"]
          format: uri
        level:
          type: string
        badge:
          type: string
        dateJoined:
          type: string
          format: date
        dateIssued:
          type: ["string", "null"]
          format: date
        passportStatus:
          type: string
          enum: [draft, active]
        active:
          type: boolean
      required:
        [id, fullName, email, phone, vocalisId, profilePhotoUrl, level, badge, dateJoined, dateIssued, passportStatus, active]
    StudentProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 120
        phone:
          type: string
          minLength: 7
          maxLength: 30
        level:
          type: string
          enum: [Level One, Level Two, Level Three, Level Four, Level Five, Level Six]
        badge:
          type: string
          enum: [Explorer, Learner, Builder, Collaborator, Leader, Changemaker, Global Citizen]
        dateJoined:
          type: string
          format: date
      required: [fullName, phone, level, badge, dateJoined]
    PhotoUploadInput:
      type: object
      properties:
        photo:
          type: string
          format: binary
      required: [photo]
...
    Passport:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        vocalisId:
          type: string
        fullName:
          type: string
        profilePhotoUrl:
          type: ["string", "null"]
          format: uri
        level:
          type: string
        badge:
          type: string
        dateJoined:
          type: string
          format: date
        dateIssued:
          type: string
          format: date
        generatedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active]
      required:
        [id, studentId, vocalisId, fullName, profilePhotoUrl, level, badge, dateJoined, dateIssued, generatedAt, status]
...