# Vocalis-Passport/lib/api-zod/src/generated/types/studentProfileUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Vocalis Passport API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentProfileUpdateBadge } from './studentProfileUpdateBadge';
import type { StudentProfileUpdateLevel } from './studentProfileUpdateLevel';

export interface StudentProfileUpdate {
  /**
     * @minLength 2
     * @maxLength 120
     */
  fullName: string;
  /**
     * @minLength 7
     * @maxLength 30
     */
  phone: string;
  level: StudentProfileUpdateLevel;
  badge: StudentProfileUpdateBadge;
  dateJoined: Date;
}